A small change...
- Put the online checking code into a separate method so that the
program can look at it during gameplay rather than waiting till game
over to do so. This saves waiting for the online top score message to
show up
- Trying to figure out how to get the "You are the top scorer!" message
to show up instantly as a consequence of this. I think I'll leave this
one for Abdel to do (or Yousuf if you're keen on doing this :P)